# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "This endpoint is used to get all users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "company_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by",
                        "name": "sort_direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Creation date",
                        "name": "creation_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Include deleted",
                        "name": "include_deleted",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Phone",
                        "name": "phone",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "This endpoint is used to create a new user",
                "parameters": [
                    {
                        "description": "User object that needs to be created",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.UserDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get authenticated users profile information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "This endpoint is used to get the authenticated users profile information using the JWT token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_domain_delivery_models_entities.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/recover/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Recover a user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "This endpoint is used to recover a user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User recovered successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/roles/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all roles of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "This endpoint is used to get all roles of a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_domain_delivery_models_entities.Role"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assign a role to a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "This endpoint is used to assign a role to a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role object that needs to be assigned",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.AssignRoleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role assigned to user successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Unassign a role from a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "This endpoint is used to unassign a role from a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role object that needs to be unassigned",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.AssignRoleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Role unassigned from user successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/sessions/{user_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Clean all sessions of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "This endpoint is used to clean all sessions of a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sessions cleaned successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MarlonG1_delivery-backend_internal_domain_delivery_models_entities.Role": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Administrator role"
                },
                "id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "github_com_MarlonG1_delivery-backend_internal_domain_delivery_models_entities.User": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "deleted_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "email_verified_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "phone": {
                    "type": "string",
                    "example": "21212828"
                },
                "phone_verified_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "profile": {
                    "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_domain_delivery_models_entities.Profile"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_domain_delivery_models_entities.UserRole"
                    }
                },
                "sessions": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_domain_delivery_models_entities.UserSession"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.AssignRoleDTO": {
            "type": "object",
            "properties": {
                "role": {
                    "description": "RoleID or RoleName of the role to be assigned to the user\n@required",
                    "type": "string",
                    "example": "admin or 3fa85f64-5717-4562-b3fc-2c963f66afa6"
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.PaginatedResponse": {
            "description": "Paginated data response with metadata about pagination",
            "type": "object",
            "properties": {
                "data": {
                    "description": "The actual data items (array of objects)"
                },
                "page": {
                    "description": "Current page number",
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "description": "Number of items per page",
                    "type": "integer",
                    "example": 20
                },
                "total_items": {
                    "description": "Total number of items across all pages",
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "description": "Total number of pages",
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.UserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email of the user to be created\n@required",
                    "type": "string",
                    "example": "example@example.com"
                },
                "full_name": {
                    "description": "Fullname of the user to be created\n@required",
                    "type": "string",
                    "example": "John Doe"
                },
                "password": {
                    "description": "Password of the user to be created\n@required",
                    "type": "string",
                    "example": "password"
                },
                "phone": {
                    "description": "Phone number of the user to be created\n@required",
                    "type": "string",
                    "example": "21212828"
                },
                "profile": {
                    "description": "Profile of the user to be created\n@required",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_dto.UserProfileDTO"
                        }
                    ]
                },
                "roles": {
                    "description": "Roles of the user to be created\n@required",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"admin\"]"
                    ]
                }
            }
        },
...
        "github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_MarlonG1_delivery-backend_internal_infrastructure_api_responser.APIError"
                }
            }
        },
...
    }
}